moved proxy spawner to zingo-proxyd, rewrote zingoproxyd.rs to use new spawner
