+++ integration-tests/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2021"
 
 
 [features]
-nym_wallet = []
-nym_server = []
+nym = []
+nym_poc = []
 
 [dependencies]
 zcash_client_backend = { workspace = true }
